move example tests closer to their modules
