+++ examples/arithmetic-expressions.test.ts
@@ -1,5 +1,5 @@
 import { assertEquals } from "@std/assert";
-import { expr } from "../examples/arithmetic-expressions.ts";
+import { expr } from "./arithmetic-expressions.ts";
 
 Deno.test("arithmetic expressions", () => {
   assertEquals(expr.parse("-1"), {
